Rename UNeXt2 package path to convnext_unet for clarity
- rename model package from models/unext2 to models/convnext_unet
- update ConvNeXtUNet imports in train.py to use the new package path
- update third-party attribution paths in README and THIRD_PARTY_NOTICES
- clarify model module documentation as UNeXt2-inspired adaptation,
  not a full Cytoland UNeXt2 reproduction
- adjust .gitignore rules so models/convnext_unet remains trackable
  despite broad models* ignore patterns

This change aligns naming with implementation scope and avoids implying
feature parity with the full UNeXt2 architecture.
