+++ THIRD_PARTY_NOTICES.md
@@ -1,8 +1,8 @@
 ## Third-Party Notices
 
-### UNeXt2 (`models/unext2/`)
+### UNeXt2-inspired ConvNeXtUNet (`models/convnext_unet/`)
 
 - Source repository: https://github.com/WayScience/virtual_stain_flow
 - Original contributor: Weishan Li
 - Upstream license: BSD 3-Clause License
-- Local status: This repository includes adapted and modified portions of the upstream implementation.
+- Local status: This repository includes adapted and modified portions of the upstream implementation and is not a full reproduction of Cytoland UNeXt2.
